backend/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
python-multipart==0.0.5
pydantic==2.4.2
pandas==2.1.3
numpy==1.26.2 